+++ src/main/java/subway/application/dto/StationResponse.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.application.dto;
 
 public class StationResponse {
     private Long id;
