+++ src/main/java/subway/ui/StationController.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
